Added additional functionality and updated classes to provide the following features:
Ask the user for a file name, and receive it from the user.  Open the file supplied by the user, and add each book to the LMS database.

Print the database to the screen, with a short message indicating that printing is taking place.

Ask the user for a barcode number to remove.  Receive the barcode, remove the corresponding book from the database, print a confirmation message that the book was deleted, and re-print the contents of the database to the screen.

Ask the user for a title number to remove.  Receive the title, remove the corresponding book from the database, print a confirmation message that the book was deleted, and re-print the contents of the database to the screen.

Ask the user for a title to check out. Check out the book, print a confirmation message, and re-print the contents of the database to the screen.

Ask the user for a title to check in. Check in the book, print a confirmation message, and re-print the contents of the database to the screen.

Still needs comments and additional formatting. 10/5/2023